Made ratvecs with 32 bit numerators/denominators to save memory and speed code. This requires removing separate instantiations of many templates which had both int and Numer forms, since now Numer = int.

+++ sources/Atlas.h
@@ -111,8 +111,10 @@ namespace atlas {
   using containers::queue;
 
   namespace arithmetic {
-    using Numer_t = long long int;
-    using Denom_t = unsigned long long int;
+    //    using Numer_t = long long int;
+    //    using Denom_t = unsigned long long int;
+    using Numer_t = int;
+    using Denom_t = unsigned int;
     template<typename I> class Rational;
     using RatNum = Rational<Numer_t>;
     using RatNumList = std::vector<RatNum>;
